# ubuntu_tank/docs/openapi_v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MentorPi Pi 5 Web Control API",
    "version": "2.0.0",
    "description": "Guarded web control, teleoperation, and lifecycle monitoring interface for MentorPi Tank on native Ubuntu 26.04 / ROS 2 Lyrical."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/version": {
      "get": {
        "summary": "Protocol and release compatibility identity",
        "description": "Returns API and protocol versions for client compatibility checks before acquiring control.",
        "responses": {
          "200": {
            "description": "Compatibility metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "summary": "Live controller and operator status",
        "responses": {
          "200": {
            "description": "Current system status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/logs": {
      "get": {
        "summary": "Bounded recent log buffer",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log lines response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/controller/stop": {
      "post": {
        "summary": "Stop controller systemd service (invalidates driving first)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControllerOperationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/control/acquire": {
      "post": {
        "summary": "Start if needed and acquire exclusive control asynchronously",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlAcquireRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Acquisition result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlAcquireResponse"
                }
              }
            }
          }
        }
      }
    },
    "/control/release": {
      "post": {
        "summary": "Relinquish operator ownership",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlReleaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Release result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlReleaseResponse"
                }
              }
            }
          }
        }
      }
    },
    "/control/arm": {
      "post": {
        "summary": "Explicitly arm after ownership and healthy preflight",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlArmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arming result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlArmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/control/stop": {
      "post": {
        "summary": "Immediate safety stop and disarm (independent of ownership)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlStopRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stop result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ControlStopResponse"
                }
              }
            }
          }
        }
      }
    },
    "/operations/{id}": {
      "get": {
        "summary": "Query asynchronous operation status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "operation_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OperatorState": {
        "type": "string",
        "enum": [
          "NO_OWNER",
          "OWNED_DISARMED",
          "ARMING",
          "ARMED_IDLE",
          "DRIVING",
          "INPUT_PAUSED",
          "FAULT"
        ]
      },
      "MotionDirection": {
        "type": "string",
        "enum": [
          "neutral",
          "forward",
          "reverse",
          "spin_left",
          "spin_right",
          "stop"
        ]
      },
      "WebControlErrorCode": {
        "type": "string",
        "enum": [
          "NOT_OWNER",
          "LEASE_EXPIRED",
          "PREFLIGHT_FAILED",
          "CONTROLLER_UNAVAILABLE",
          "DEPLOYMENT_BUSY",
          "INVALID_EPOCH",
          "CHALLENGE_REUSED",
          "CHALLENGE_EXPIRED",
          "SEQUENCE_OUT_OF_ORDER",
          "MAX_HOLD_EXCEEDED",
          "STALE_TELEMETRY",
          "INVALID_STATE",
          "INPUT_CONFLICT",
          "UNAUTHORIZED",
          "INVALID_PAYLOAD",
          "TRACKS_NOT_RAISED",
          "OPERATION_FAILED",
          "TIMEOUT",
          "STALE_TRANSACTION",
          "INCOMPATIBLE_PROTOCOL"
        ]
      },
      "ControllerServiceState": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "failed",
          "unknown"
        ]
      },
      "VersionResponse": {
        "type": "object",
        "required": [
          "protocol_version",
          "api_version",
          "schema_version",
          "release_id",
          "supported_protocols"
        ],
        "properties": {
          "protocol_version": {
            "type": "string"
          },
          "api_version": {
            "type": "string"
          },
          "schema_version": {
            "type": "integer"
          },
          "release_id": {
            "type": "string"
          },
          "supported_protocols": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "required": [
          "service_state",
          "operator_state",
          "guard_armed",
          "disarm_pending",
          "linear_speed",
          "angular_speed",
          "limits",
          "freshness",
          "release_id",
          "protocol_version"
        ],
        "properties": {
          "service_state": {
            "$ref": "#/components/schemas/ControllerServiceState"
          },
          "operator_state": {
            "$ref": "#/components/schemas/OperatorState"
          },
          "active_owner": {
            "type": "string",
            "nullable": true
          },
          "current_epoch": {
            "type": "integer",
            "nullable": true
          },
          "guard_armed": {
            "type": "boolean",
            "nullable": true
          },
          "disarm_pending": {
            "type": "boolean"
          },
          "battery_voltage": {
            "type": "number",
            "nullable": true
          },
          "linear_speed": {
            "type": "number"
          },
          "angular_speed": {
            "type": "number"
          },
          "limits": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "freshness": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "nullable": true
            }
          },
          "last_fault": {
            "type": "string",
            "nullable": true
          },
          "release_id": {
            "type": "string"
          },
          "protocol_version": {
            "type": "string"
          },
          "input_generation": {
            "type": "integer"
          },
          "pause_reason": {
            "type": "string",
            "nullable": true
          },
          "recovery_ready": {
            "type": "boolean"
          }
        }
      },
      "LogsResponse": {
        "type": "object",
        "required": [
          "lines",
          "total_lines",
          "total_bytes"
        ],
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total_lines": {
            "type": "integer"
          },
          "total_bytes": {
            "type": "integer"
          }
        }
      },
      "ControllerOperationRequest": {
        "type": "object",
        "required": [
          "request_id",
          "action"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": [
              "start",
              "stop"
            ]
          }
        }
      },
      "OperationStatusResponse": {
        "type": "object",
        "required": [
          "operation_id",
          "status"
        ],
        "properties": {
          "operation_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "failed"
            ]
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "epoch": {
            "type": "integer",
            "nullable": true
          },
          "bind_token": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ControlAcquireRequest": {
        "type": "object",
        "required": [
          "request_id",
          "operator_id",
          "protocol_version"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "operator_id": {
            "type": "string"
          },
          "protocol_version": {
            "type": "string",
            "enum": [
              "2.0.0"
            ]
          }
        }
      },
      "ControlAcquireResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "epoch": {
            "type": "integer",
            "nullable": true
          },
          "bind_token": {
            "type": "string",
            "nullable": true
          },
          "active_owner": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/WebControlErrorCode",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "operation_id": {
            "type": "string",
            "nullable": true
          },
          "operation_token": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ControlReleaseRequest": {
        "type": "object",
        "required": [
          "request_id"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "epoch": {
            "type": "integer"
          },
          "operation_id": {
            "type": "string"
          },
          "operation_token": {
            "type": "string"
          }
        }
      },
      "ControlReleaseResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/WebControlErrorCode",
            "nullable": true
          }
        }
      },
      "ControlArmRequest": {
        "type": "object",
        "required": [
          "request_id",
          "epoch"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "epoch": {
            "type": "integer"
          }
        },
        "additionalProperties": false
      },
      "ControlArmResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/WebControlErrorCode",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ControlStopRequest": {
        "type": "object",
        "required": [
          "request_id"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "epoch": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "ControlStopResponse": {
        "type": "object",
        "required": [
          "success",
          "disarmed"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "disarmed": {
            "type": "boolean"
          }
        }
      },
      "Challenge": {
        "type": "object",
        "required": [
          "token",
          "epoch",
          "deadline_monotonic_ns",
          "issued_monotonic_ns",
          "input_generation"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "epoch": {
            "type": "integer"
          },
          "deadline_monotonic_ns": {
            "type": "integer"
          },
          "issued_monotonic_ns": {
            "type": "integer"
          },
          "input_generation": {
            "type": "integer",
            "minimum": 0
          },
          "recovery_required": {
            "type": "boolean"
          }
        }
      },
      "ChallengeResponse": {
        "type": "object",
        "required": [
          "token",
          "epoch",
          "sequence",
          "direction",
          "input_generation"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "epoch": {
            "type": "integer"
          },
          "sequence": {
            "type": "integer"
          },
          "direction": {
            "$ref": "#/components/schemas/MotionDirection"
          },
          "client_timestamp_ms": {
            "type": "integer",
            "nullable": true
          },
          "input_generation": {
            "type": "integer",
            "minimum": 0
          }
        }
      }
    }
  }
}